# openapi.yaml
openapi: 3.0.0
info:
  x-skaffolder-id-project: 61a8d1825db62a1b83fc156c
  title: schoolmanagement
  description: You can obtain JWT token from login API
  version: 0.1
servers:
  - url: http://localhost:3000/api
    description: Local server for testing
paths:
...
  "/student/{id}":
    delete:
      x-skaffolder-id: 61a8d5a75db62a1b83fc1607
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61a8d3b65db62a1b83fc15e4
      x-skaffolder-resource: student
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61a8d5a75db62a1b83fc15ff
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61a8d3b65db62a1b83fc15e4
      x-skaffolder-resource: student
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
    post:
      x-skaffolder-id: 61a8d5a75db62a1b83fc15fd
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61a8d3b65db62a1b83fc15e4
      x-skaffolder-resource: student
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - student
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/student"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/student"
  "/student/findBy_courses/{key}":
    get:
      x-skaffolder-id: 61a8d5eb5db62a1b83fc161f
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 61a8d3b65db62a1b83fc15e4
      x-skaffolder-resource: student
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - student
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/student"
...
  "/teacher/{id}":
    delete:
      x-skaffolder-id: 61a8d5d85db62a1b83fc1612
      x-skaffolder-name: delete
      x-skaffolder-id-resource: 61a8d4055db62a1b83fc15ea
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: delete
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION delete
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION delete
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
    get:
      x-skaffolder-id: 61a8d5d85db62a1b83fc160b
      x-skaffolder-name: get
      x-skaffolder-id-resource: 61a8d4055db62a1b83fc15ea
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: get
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION get
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION get
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id resource
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
    post:
      x-skaffolder-id: 61a8d5d85db62a1b83fc160d
      x-skaffolder-name: update
      x-skaffolder-id-resource: 61a8d4055db62a1b83fc15ea
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: update
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION update
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /{id}
      x-skaffolder-roles: []
      summary: CRUD ACTION update
      tags:
        - teacher
      parameters:
        - name: id
          x-skaffolder-type: ObjectId
          in: path
          description: Id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/teacher"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/teacher"
  "/teacher/findBy_courses/{key}":
    get:
      x-skaffolder-id: 61a8d5eb5db62a1b83fc161d
      x-skaffolder-name: findBy_courses
      x-skaffolder-id-resource: 61a8d4055db62a1b83fc15ea
      x-skaffolder-resource: teacher
      x-skaffolder-crudAction: findBy_courses
      x-skaffolder-crudType: 
      x-skaffolder-description: CRUD ACTION findBy_courses
      x-skaffolder-returnDesc: 
      x-skaffolder-returnType: 
      x-skaffolder-url: /findBy_courses/{key}
      x-skaffolder-roles: []
      summary: CRUD ACTION findBy_courses
      tags:
        - teacher
      parameters:
        - name: key
          x-skaffolder-type: Objectid
          in: path
          description: Id of model to search for
          required: true
          schema:
            type: Objectid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/teacher"
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    student:
      x-skaffolder-id: 61a8d3b65db62a1b83fc15e4
      x-skaffolder-id-db: 61a8d1825db62a1b83fc156d
      x-skaffolder-id-entity: 61a8d3b65db62a1b83fc15e1
      x-skaffolder-url: /student
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 61a8d3b65db62a1b83fc15e2
          x-skaffolder-type: String
        surename:
          type: string
          x-skaffolder-id-attr: 61a8d3b65db62a1b83fc15e3
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 61a8d5795db62a1b83fc15fb
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 61a8d3b65db62a1b83fc15e1
          x-skaffolder-ent2: 61a8d42d5db62a1b83fc15ed
      required:
        - _id
    teacher:
      x-skaffolder-id: 61a8d4055db62a1b83fc15ea
      x-skaffolder-id-db: 61a8d1825db62a1b83fc156d
      x-skaffolder-id-entity: 61a8d4055db62a1b83fc15e7
      x-skaffolder-url: /teacher
      properties:
        _id:
          type: string
          x-skaffolder-required: true
        name:
          type: string
          x-skaffolder-id-attr: 61a8d4055db62a1b83fc15e8
          x-skaffolder-type: String
        surename:
          type: string
          x-skaffolder-id-attr: 61a8d4055db62a1b83fc15e9
          x-skaffolder-type: String
      x-skaffolder-relations:
        _courses:
          x-skaffolder-id: 61a8d5345db62a1b83fc15fa
          x-skaffolder-type: m:m
          x-skaffolder-ent1: 61a8d4055db62a1b83fc15e7
          x-skaffolder-ent2: 61a8d42d5db62a1b83fc15ed
      required:
        - _id
...